FINAL_resources/app_demo/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
Pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.1.78
werkzeug==3.0.0
